Added option to specify gene table
For some genomes the refGene annotations are quite sparse, resulting in many regions being annotated with genes that are actually not very nearby. In these cases it may be beneficial to be able to specify a different UCSC table to use for gene annotations, for instance one with more gene predictions. I added this option for myself, but this may be beneficial for others as well. I set the default to "refGene", meaning that when the new table argument is ignored, comb-p should run as usual.